refactor(http): share feed to_xml via impl_feed_to_xml macro
AtomFeed and Rss2Feed had an identical to_xml: drain into_stream_writer() into
a Vec<u8>. Extract it into an impl_feed_to_xml! macro defined in the rss module
(in textual scope for both submodules).

Behavior unchanged; rss tests green, clippy clean.
